requirements.txt
aiohttp==3.12.7
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==4.2.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
curl_cffi==0.10.0
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.12
feedparser==6.0.11
Flask==3.1.1
fonttools==4.57.0
frozendict==2.4.6
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
investpy==1.0.8
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mplfinance==0.12.10b0
multitasking==0.0.11
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pandas==2.2.1
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
praw==7.8.1
prawcore==2.4.0
protobuf==6.30.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
textblob==0.19.0
tornado==6.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
update-checker==0.18.0
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.61
